apps/api/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
celery>=5.4.0
redis>=5.0.4
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
pydantic-settings>=2.3.0
pdfplumber>=0.11.0
pydantic>=2.7.0
faster-whisper>=1.0.0
pyannote.audio>=3.1.1
openai>=1.0.0
